+++ meteor-typescript-compiler.ts
@@ -1,7 +1,60 @@
 import * as ts from "typescript";
 import * as crypto from "crypto";
-import { CompilerCache } from "compiler-cache";
-import { setTraceEnabled, error, info, trace } from "compiler-console";
+import { bold, dim, reset } from "chalk";
+
+/**
+ * compiler-console (could not figure out how to load from separate file/module)
+ * only got errors of type "packages/modules-runtime.js:222:12: Cannot find module 'compiler-console'"
+ */
+
+let traceEnabled = false;
+
+export function setTraceEnabled(enabled: boolean) {
+  traceEnabled = enabled;
+}
+
+export function error(msg: string, ...other: string[]) {
+  process.stderr.write(bold.red(msg) + reset(other.join(" ")) + "\n");
+}
+
+export function info(msg: string) {
+  process.stdout.write(bold.green(msg) + dim(" ") + "\n");
+}
+
+export function trace(msg: string) {
+  if (traceEnabled) {
+    process.stdout.write(dim(msg) + dim(" ") + "\n");
+  }
+}
+
+/**
+ * compiler-cache (could not figure out how to load from separate file/module)
+ */
+interface JavascriptData {
+  source: string;
+  fileName: string;
+}
+
+interface CacheData {
+  javascript: JavascriptData;
+  sourceMapJson: string | undefined;
+}
+/**
+ * Caches output from typescript on disk
+ */
+export class CompilerCache {
+  constructor(private cacheRoot: string) {}
+
+  public addJavascript(sourceFilePath: string, data: JavascriptData) {
+    info(`Caching javascript for ${sourceFilePath}`);
+  }
+  public addSourceMap(sourceFilePath: string, data: string) {
+    info(`Caching sourceMap for ${sourceFilePath}`);
+  }
+  public get(sourceFilePath: string): CacheData | undefined {
+    return undefined;
+  }
+}
 
 interface LocalEmitResult {
   fileName: string;
